Refactor UI components and variants, improve type safety and code organization
